requirements.txt
pydantic>=2.0
chromadb>=0.4.0
openai>=1.0.0
python-dotenv>=1.0.0
pdfplumber>=0.11.0
streamlit>=1.30.0
